requirements.txt
openai==1.43.0
Flask==3.0.3
python-dotenv==1.0.1
requests==2.32.3
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
gunicorn==22.0.0